requirements.txt
prometheus-client==0.22.1
python-dotenv==1.1.1
python-kasa==0.10.2
